feat(ui5-textarea): migrate to Typescript Related to: #4337
